move navigation_bar_destination object on entity folder
move the object to entity folder for mantain consistency with folders
structure of project
