+++ yomu_no_ikiru/lib/core/common/entities/navigation_bar_destination.dart
